\docType{class}
\name{AssociationRulesAlgorithm}
\alias{AssociationRulesAlgorithm}
\title{Association Rules Algorithm}
\description{
Class inheriting of KeelAlgorithm, to common methods for all KEEL Association Rules Algorithms.
The specific association rules algorithms must inherit of this class.
}
\keyword{association rules}

