\docType{class}
\name{ClassificationAlgorithm}
\alias{ClassificationAlgorithm}
\title{Classification Algorithm}
\description{
Class inheriting of KeelAlgorithm, to common methods for all KEEL Classification Algorithms.
The specific classification algorithms must inherit of this class.
}
\keyword{classification}

