\docType{class}
\name{ClassificationResults}
\alias{ClassificationResults}
\title{Classification Results}
\description{
Class to calculate and store some results for a ClassificationAlgorithm.
It receives as parameter the prediction of a classification algorithm as a data.frame object.
}
\keyword{classification}

