\docType{methods}
\name{getKeelDatasetList}
\alias{getKeelDatasetList}
\title{Get Keel Dataset List}
\description{
Method for knowing which datasets from the KEEL Dataset Repository are availabre at RKEEL.
}
\usage{
getKeelDatasetList()
}
\value{
Returns a list with the data names.
}
\keyword{utils}

