\docType{class}
\name{ImbalancedClassificationAlgorithm}
\alias{ImbalancedClassificationAlgorithm}
\title{ImbalancedClassification Algorithm}
\description{
Class inheriting of ClassificationAlgorithm, to common methods for all KEEL Imbalanced Classification Algorithms.
The specific imbalanced-classification algorithms must inherit of this class.
}
\keyword{classification}

