\docType{methods}
\name{read.keel}
\alias{read.keel}
\title{Read keel dataset}
\description{
Method for read datasets in .dat KEEL format
}
\usage{
read.keel(file)
}
\arguments{
\item{file}{File containing the dataset to be read. It must be in KEEL .dat format.}
}
\value{
Returns a data.frame object with the dataset
}
\keyword{utils}

