\docType{methods}
\name{writeDatFromDataframes}
\alias{writeDatFromDataframes}
\title{Write .dat from data.frames}
\description{
Method for writing both train and test .dat dataset files in KEEL format.
}
\usage{
writeDatFromDataframes(trainData, testData,
   trainFileName, testFileName)
}
\arguments{
\item{trainData}{Train data as data.frame object}
\item{testData}{Test data as data.frame object}
\item{trainFileName}{String with the file name to store the train dataset}
\item{testFileName}{String with the file name to store the test dataset}
}
\keyword{utils}

